requirements.txt
iniconfig==2.1.0
MarkupSafe==3.0.2
packaging==25.0
pathlib==1.0.1
pluggy==1.5.0
pytest==8.3.5
ruamel.yaml==0.18.10
